refactor : ClinicPage 리팩토링 (#96)
컴포넌트 최상위 ClinicCientPage에서는 전체적인 구조가 보이도록 설정
ScheduleList 내부에서 Header를 별도로 분리하여 로직 가독성 높임
ActionTag 배럴파일로 가져오도록 설정
