+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './actionTag/ActionTag'
+export { default as ActionTag } from './actionTag/ActionTag'
 export * from './bottomNav/BottomNav'
 export * from './bottomSheet/BottomSheet'
 export * from './button/Button'
