+++ src/features/clinic/ui/ClinicClientPage.tsx
@@ -1,9 +1,6 @@
 'use client'
 
-import { useParams, useRouter } from 'next/navigation'
-
-import { BottomNav, Calendar, Label, Profile } from '@/components'
-import ActionTag from '@/components/actionTag/ActionTag'
+import { BottomNav, Calendar, Profile } from '@/components'
 import { ScheduleList } from '@/features'
 import { useUserStore } from '@/store'
 
@@ -11,13 +8,7 @@ import { useScheduleList } from '../service/useScheduleList'
 
 export const ClinicClientPage = () => {
   const { user } = useUserStore()
-  const { userId } = useParams<{ userId: string }>()
-  const router = useRouter()
   const scheduleList = useScheduleList()
-
-  const handleClickPlusButton = () => {
-    router.push(`/clinic/${userId}/write`)
-  }
 
   return (
     <>
@@ -31,10 +22,6 @@ export const ClinicClientPage = () => {
         </section>
 
         <section className="mt-7">
-          <div className="flex-between mb-3">
-            <Label icon="clinic-label">병원 내원 일정</Label>
-            <ActionTag.Plus label="추가" primary onClick={handleClickPlusButton} />
-          </div>
           <ScheduleList scheduleList={scheduleList} />
         </section>
       </main>
