+++ src/features/clinic/ui/ScheduleItem.tsx
@@ -1,6 +1,7 @@
 'use client'
 
-import { useParams, useRouter } from 'next/navigation'
+import Link from 'next/link'
+import { useParams } from 'next/navigation'
 import dayjs from 'dayjs'
 
 import { Icon, Tag } from '@/components'
@@ -15,36 +16,29 @@ type ScheduleItemProps = {
 }
 
 export const ScheduleItem = ({ item }: ScheduleItemProps) => {
-  const router = useRouter()
   const { userId } = useParams<{ userId: string }>()
   const { medicalId, hospitalDate, hospitalName, medicalPart } = item
 
   const time = dayjs(hospitalDate).format('HH:mm')
   const date = dayjs(hospitalDate).format('M월 D일(dd) A HH:mm')
 
-  const handleClickItem = () => {
-    router.push(`/clinic/${userId}/info/${medicalId}`)
-  }
-
   return (
-    <div className="rounded-xl bg-mint-0 p-4">
-      <button className="size-full text-left" onClick={handleClickItem}>
-        <div className="flex-between">
-          <div className="flex-col">
-            <p className="subtitle-B">
-              {time} {hospitalName}
-            </p>
-            <p className="headline-M text-gray-6">{date}</p>
-          </div>
-          <Icon name="check-no" size={36} />
+    <Link href={`/clinic/${userId}/info/${medicalId}`} className="rounded-xl bg-mint-0 p-4">
+      <div className="flex-between">
+        <div className="flex-col">
+          <p className="subtitle-B">
+            {time} {hospitalName}
+          </p>
+          <p className="headline-M text-gray-6">{date}</p>
         </div>
+        <Icon name="check-no" size={36} />
+      </div>
 
-        <div className="mt-8 flex flex-wrap gap-2">
-          {medicalPart.map((part) => (
-            <Tag key={part} label={part} />
-          ))}
-        </div>
-      </button>
-    </div>
+      <div className="mt-8 flex flex-wrap gap-2">
+        {medicalPart.map((part) => (
+          <Tag key={part} label={part} />
+        ))}
+      </div>
+    </Link>
   )
 }
